+++ src/lib/hooks/useBusinessUnits/index.ts
@@ -1,23 +1,46 @@
 import { useCallback } from 'react';
 import { sdk } from '@/sdk';
 import useSWR from 'swr';
-import { BusinessUnit } from '@shared/types/business-unit/BusinessUnit';
 import { Associate } from '@/types/entity/associate';
-import { Account } from '@shared/types/account/Account';
-import { Store } from '@shared/types/store/Store';
+import { Address } from '@shared/types/account';
 
 const useBusinessUnits = () => {
   const { data, mutate: mutateBusinessUnits } = useSWR('/action/business-unit/getBusinessUnits', () =>
-    sdk.callAction<BusinessUnit[]>({ actionName: 'business-unit/getBusinessUnits' }),
+    sdk.composableCommerce.businessUnit.getBusinessUnits(),
   );
 
   const addBusinessUnit = useCallback(
-    async (payload: { account: Account; store: Store }) => {
-      const response = await sdk.callAction<BusinessUnit>({
-        actionName: 'business-unit/create',
-        payload,
-        query: { storeKey: payload.store.key as string },
+    async (payload: {
+      account: {
+        accountId: string;
+        email: string;
+        companyName: string;
+      };
+      storeId: string;
+    }) => {
+      const response = await sdk.composableCommerce.businessUnit.createBusinessUnit({
+        account: payload.account,
+        store: {
+          storeId: payload.storeId,
+        },
       });
+
+      mutateBusinessUnits();
+
+      return response.isError ? {} : response.data;
+    },
+    [mutateBusinessUnits],
+  );
+
+  const updateBusinessUnit = useCallback(
+    async (payload: { name: string; contactEmail: string; key: string }) => {
+      const response = await sdk.composableCommerce.businessUnit.updateBusinessUnit(
+        {
+          name: payload.name,
+          contactEmail: payload.contactEmail,
+        },
+        { key: payload.key },
+      );
 
       mutateBusinessUnits();
 
@@ -28,7 +51,7 @@ const useBusinessUnits = () => {
 
   const removeBusinessUnit = useCallback(
     async (key: string) => {
-      const response = await sdk.callAction<BusinessUnit>({ actionName: 'business-unit/remove', query: { key } });
+      const response = await sdk.composableCommerce.businessUnit.removeBusinessUnit({ key });
 
       mutateBusinessUnits();
 
@@ -39,11 +62,10 @@ const useBusinessUnits = () => {
 
   const addAssociate = useCallback(
     async ({ email, role, businessUnit }: Partial<Associate> & { businessUnit: string }) => {
-      const response = await sdk.callAction<BusinessUnit>({
-        actionName: 'business-unit/addAssociate',
-        payload: { email, roleKeys: [role] },
-        query: { key: businessUnit },
-      });
+      const response = await sdk.composableCommerce.businessUnit.addAssociate(
+        { email: email as string, roleKeys: [role as string] },
+        { key: businessUnit },
+      );
 
       mutateBusinessUnits();
 
@@ -54,11 +76,10 @@ const useBusinessUnits = () => {
 
   const updateAssociate = useCallback(
     async ({ id, role, businessUnit }: Partial<Associate> & { businessUnit: string }) => {
-      const response = await sdk.callAction<BusinessUnit>({
-        actionName: 'business-unit/updateAssociate',
-        payload: { accountId: id, roleKeys: [role] },
-        query: { key: businessUnit },
-      });
+      const response = await sdk.composableCommerce.businessUnit.updateAssociate(
+        { accountId: id as string, roleKeys: [role as string] },
+        { key: businessUnit },
+      );
 
       mutateBusinessUnits();
 
@@ -69,11 +90,43 @@ const useBusinessUnits = () => {
 
   const removeAssociate = useCallback(
     async ({ id, businessUnit }: Partial<Associate> & { businessUnit: string }) => {
-      const response = await sdk.callAction<BusinessUnit>({
-        actionName: 'business-unit/removeAssociate',
-        payload: { accountId: id },
-        query: { key: businessUnit },
-      });
+      const response = await sdk.composableCommerce.businessUnit.removeAssociate(
+        { accountId: id ?? '' },
+        { key: businessUnit },
+      );
+
+      mutateBusinessUnits();
+
+      return response.isError ? {} : response.data;
+    },
+    [mutateBusinessUnits],
+  );
+
+  const addAddress = useCallback(
+    async ({ businessUnit, ...address }: Address & { businessUnit: string }) => {
+      const response = await sdk.composableCommerce.businessUnit.addAddress({ address }, { key: businessUnit });
+
+      mutateBusinessUnits();
+
+      return response.isError ? {} : response.data;
+    },
+    [mutateBusinessUnits],
+  );
+
+  const updateAddress = useCallback(
+    async ({ businessUnit, ...address }: Address & { businessUnit: string }) => {
+      const response = await sdk.composableCommerce.businessUnit.updateAddress({ address }, { key: businessUnit });
+
+      mutateBusinessUnits();
+
+      return response.isError ? {} : response.data;
+    },
+    [mutateBusinessUnits],
+  );
+
+  const removeAddress = useCallback(
+    async ({ businessUnit, addressId }: { addressId: string; businessUnit: string }) => {
+      const response = await sdk.composableCommerce.businessUnit.removeAddress({ addressId }, { key: businessUnit });
 
       mutateBusinessUnits();
 
@@ -86,10 +139,14 @@ const useBusinessUnits = () => {
     businessUnits: data?.isError ? [] : data?.data ?? [],
     defaultBusinessUnit: data?.isError ? undefined : data?.data[0],
     addBusinessUnit,
+    updateBusinessUnit,
     removeBusinessUnit,
     addAssociate,
     updateAssociate,
     removeAssociate,
+    addAddress,
+    updateAddress,
+    removeAddress,
   };
 };