+++ src/lib/tastics/index.tsx
@@ -22,6 +22,8 @@ import AddressesTastic from './addresses';
 import PurchaseListDetailTastic from './purchase-list-detail';
 import OrderDetailTastic from './order-detail';
 import QuoteDetailTastic from './quote-detail';
+import CheckoutTastic from './checkout';
+import ThankYouTastic from './thank-you';
 
 const tastics = {
   'commercetools/ui/content/hero-tile': HeroTileTastic,
@@ -47,6 +49,8 @@ const tastics = {
   'commercetools/ui/purchase-list-detail': PurchaseListDetailTastic,
   'commercetools/ui/order-detail': OrderDetailTastic,
   'commercetools/ui/quote-detail': QuoteDetailTastic,
+  'commercetools/ui/checkout': CheckoutTastic,
+  'commercetools/ui/thank-you': ThankYouTastic,
 } as TasticRegistry;
 
 export default tastics;