+++ src/lib/tastics/product-list/index.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { redirect } from 'next/navigation';
 import { DataSource } from '@/types/lib/datasources';
 import { sdk } from '@/sdk';
-import { Result } from '@shared/types/product/Result';
 import { Category } from '@shared/types/product/Category';
 import { TasticProps } from '../types';
 import { CategoryConfiguration, DataSourceProps, Props } from './types';
@@ -11,14 +10,8 @@ import ProductListViewModel from './components/product-list-view-model';
 const ProductListTastic = async ({ data, searchParams }: TasticProps<DataSource<DataSourceProps> & Props>) => {
   if (!data.data?.dataSource?.items?.length) return redirect('/404');
 
-  const flatCategoriesResponse = await sdk.callAction<Result>({
-    actionName: 'product/queryCategories',
-  });
-
-  const categoriesResponse = await sdk.callAction<Result>({
-    actionName: 'product/queryCategories',
-    query: { format: 'tree' },
-  });
+  const flatCategoriesResponse = await sdk.composableCommerce.product.queryCategories();
+  const categoriesResponse = await sdk.composableCommerce.product.queryCategories({ format: 'tree' });
 
   const flatCategories = flatCategoriesResponse.isError ? [] : (flatCategoriesResponse.data.items as Category[]);
   const categories = categoriesResponse.isError ? [] : (categoriesResponse.data.items as Category[]);
@@ -26,6 +19,7 @@ const ProductListTastic = async ({ data, searchParams }: TasticProps<DataSource<
   const slug = data.data.dataSource?.category?.split('/')?.at(-1) ?? '';
 
   const category = flatCategories.find((c) => c.slug === slug);
+  const treeCategory = categories.find((c) => c.slug === slug);
 
   const isRootCategory = category?.depth === 0;
 
@@ -33,13 +27,14 @@ const ProductListTastic = async ({ data, searchParams }: TasticProps<DataSource<
     .find((key) => key.startsWith('sortAttributes'))
     ?.match(/sortAttributes\[0\]\[(.+)\]/)?.[1];
 
-  const categoryConfiguration = data.categoryConfiguration.reduce(
+  const categoryConfiguration = (data.categoryConfiguration ?? []).reduce(
     (acc, curr) => ({ ...acc, [curr.key]: curr }),
     {} as Record<string, CategoryConfiguration>,
   );
 
-  if ((!isRootCategory || (isRootCategory && searchParams.view)) && !sortValueExists)
-    return redirect('?sortAttributes[0][price]=asc&view=1');
+  const isIntermediaryPage = isRootCategory && !searchParams.view && (treeCategory?.subCategories.length ?? 0) >= 3;
+
+  if (!isIntermediaryPage && !sortValueExists) return redirect('?sortAttributes[0][price]=asc&view=1');
 
   return (
     <ProductListViewModel
@@ -47,6 +42,7 @@ const ProductListTastic = async ({ data, searchParams }: TasticProps<DataSource<
       categoryConfiguration={categoryConfiguration}
       categories={flatCategories}
       category={categories.find((c) => c.slug === slug) ?? category}
+      displayIntermediaryPage={isIntermediaryPage}
     />
   );
 };