+++ src/lib/tastics/thank-you/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ThankYouTastic = () => {
+  return <div>ThankYouTastic</div>;
+};
+
+export default ThankYouTastic;